graph domain settings

+++ src/routes/app/graph/[graph]/settings/DomainRow.svelte
@@ -1,13 +1,21 @@
 
 <script lang="ts">
 
+	// External dependencies
+	import { onMount } from 'svelte'
+	import { writable } from 'svelte/store'
+
 	// Internal dependencies
+	import { domains } from './stores'
 	import { DomainController, SortOption } from '$scripts/controllers'
+	import { ValidationData } from '$scripts/validation'
+
+	import type { DropdownOption } from '$scripts/types'
 
 	// Components
-	import IconButton from '$components/buttons/IconButton.svelte'
-	import Textfield from '$components/forms/Textfield.svelte'
-	import Dropdown from '$components/forms/Dropdown.svelte'
+	import IconButton from '$components/IconButton.svelte'
+	import Textfield from '$components/Textfield.svelte'
+	import Dropdown from '$components/Dropdown.svelte'
 	import Validation from '$components/Validation.svelte'
 
 	// Assets
@@ -15,8 +23,19 @@
 
 	// Exports
 	export let domain: DomainController
-	export let sortmode: (options: number) => void
+	export let updateSortmode: (options: number) => void
 	export let update: () => void
+
+	// Stores
+	const style_options = writable<DropdownOption<string>[] | undefined>()
+	const validation = writable<ValidationData | undefined>()
+	const color = writable<string | undefined>()
+
+	onMount(async () => {
+		domains.subscribe(async () => style_options.set(await domain.getStyleOptions()))
+		domains.subscribe(async () => validation.set(await domain.validate()))
+		domains.subscribe(async () => color.set(await domain.getColor()))
+	})
 
 </script>
 
@@ -24,53 +43,48 @@
 <!-- Markdown -->
 
 
-<div class="row" id={domain.uuid}>
-	{#await domain.validate() then validate}
-		<Validation short data={validate} />
-	{/await}
+{#if $style_options !== undefined && $validation !== undefined && $color !== undefined}
+	<div class="domain row" id={domain.uuid}>
+		<Validation compact data={$validation} />
 
-	{#await domain.getIndex() then index}
-		<span> {index + 1} </span>
-	{/await}
-
-	<IconButton scale
-		src={trashIcon}
-		on:click={async () => {
-			await domain.delete()
-			sortmode(SortOption.domains)
-			update()
-		}}
+		<IconButton scale
+			src={trashIcon}
+			on:click={async () => {
+				await domain.delete()
+				updateSortmode(SortOption.domains)
+				update()
+			}}
 		/>
 
-	<Textfield
-		id="name"
-		placeholder="Domain Name"
-		bind:value={domain.name}
-		on:change={async () => {
-			await domain.save()
-			sortmode(SortOption.domains)
-			update()
-		}}
+		<span> {domain.index} </span>
+
+		<Textfield
+			id="name"
+			placeholder="Domain Name"
+			bind:value={domain.name}
+			on:change={async () => await domain.save()}
+			on:input={() => {
+				updateSortmode(SortOption.domains)
+				update()
+			}}
 		/>
 
-	{#await domain.getStyleOptions() then style_options}
 		<Dropdown
 			id="style"
 			placeholder="Domain Style"
-			options={style_options}
+			options={$style_options}
 			bind:value={domain.style}
 			on:change={async () => {
 				await domain.save()
-				sortmode(SortOption.domains)
+				updateSortmode(SortOption.domains)
 				update()
 			}}
-			/>
-	{/await}
+		/>
 
-	{#await domain.getColor() then color}
-		<span class="preview" style:background-color={color} />
-	{/await}
-</div>
+
+		<span class="preview" style:background-color={$color} />
+	</div>
+{/if}
 
 
 <!-- Styles -->
@@ -78,31 +92,6 @@
 
 <style lang="sass">
 
-	@use "$styles/variables.sass" as *
-	@use "$styles/palette.sass" as *
-
-	$icon-width: calc($input-icon-size + 2 * $input-icon-padding)
-
-	.row
-		display: grid
-		grid-template: "validation id delete left right right-preview" auto / $icon-width $icon-width $icon-width 1fr 1fr $icon-width
-		place-items: center center
-		gap: $form-small-gap
-
-		padding-right: calc( $form-small-gap + $icon-width )
-
-		.preview
-			width: $input-icon-size
-			height: $input-icon-size
-
-		.header
-			display: flex
-			flex-flow: row nowrap
-			align-content: center
-			justify-content: right
-			width: 100%
-
-			span
-				flex: 1
+	@import './domain_styles.sass'
 
 </style>
